+++ lizard_portal/models.py
@@ -63,6 +63,9 @@ class ConfigurationToValidate(models.Model):
         help_text='Type of the configuration',
         max_length=24)
     # allowed values for config_type are 'waterbalans', 'esf[n]' for n > 0
+    name = models.CharField(
+        help_text='Configuration name supplied with the configuration',
+        null=True, blank=True, max_length=256)
     user = models.CharField(
         help_text='User name supplied with the configuration',
         max_length=48)
@@ -105,7 +108,7 @@ class ConfigurationToValidate(models.Model):
                     self.data_set = self.db.data_sets.get(name__iexact=value)
                 except DataSet.DoesNotExist:
                     logger.warning("Unable to find the data set with name '%s'", value)
-            elif key in ['file_path', 'config_type', 'date', 'user']:
+            elif key in ['file_path', 'config_type', 'date', 'user', 'name']:
                 setattr(self, key, value)
             else:
                 meta_info += '%s: %s; ' % (key, value)
@@ -120,6 +123,7 @@ class ConfigurationToValidate(models.Model):
             'date':       self.date,
             'action':     self.get_action_display(),
             'action_log': self.action_log,
+            'name':       self.name,
             }
 
     @property